+++ src/icalendar/timezone/__init__.py
@@ -8,6 +8,7 @@ def use_pytz():
     """Use pytz as the implementation that looks up and creates timezones."""
     tzp.use_pytz()
 
+
 def use_zoneinfo():
     """Use zoneinfo as the implementation that looks up and creates timezones."""
     tzp.use_zoneinfo()